Refactor: using the split-out auth package (#2355)
* Vendoring 0.1.0 of github.com/hashicorp/go-azure-helpers

* refactor: moving auth over to the new package

* refactor: moving resource provider registration over to the split-out sdk

* refactor: passing through to the new sdk

* vendoring the `2017-03-09` profile

* Updating to `0.1.1` of github.com/hashicorp/go-azure-helpers

* removing the unused sdk
